% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIriskjobs.R
\docType{data}
\name{AIriskjobs}
\alias{AIriskjobs}
\title{Jobs that are affected by AI and Automation}
\format{
\subsection{'AIriskjobs'}{

A data frame with 500 rows and 10 columns:
\describe{
\item{Job_Title}{The specific job title being analyzed}
\item{Industry}{The industry or sector to which the job belongs}
\item{Company_Size}{Size of that company (small, medium, large)}
\item{Location}{Location of the job}
\item{AI_adoption_level}{The level of AI adoption in that sector (low, medium, high)}
\item{Automation_Risk}{The likelihood that the job will be automated (low, medium, high)}
\item{Required_Skills}{Skills necessary for the job role}
\item{Salary_USD}{Average Salary of that job in US Dollars}
\item{Remote_Friendly}{Indicates whether the job allows for remote work}
\item{Job_Growth_projection}{The projected growth or decline of jobs in this sector}
}
}
}
\source{
\url{https://www.kaggle.com/datasets/uom190346a/ai-powered-job-market-insights}
}
\usage{
AIriskjobs
}
\description{
A complete list of jobs that are affected by AI and Automation.
}
\keyword{datasets}
